Move SingleOrVec tests from http.rs to single_or_vec.rs for better organization
